feat: VW_UnitAttackRange 내부에서 Sprite의 크기를 조절하는 기능 구현

+++ supermaker_assignment/Assets/Scripts/Unit/UnitGroup.cs
@@ -89,10 +89,10 @@ namespace Unit
         /// 현재 그룹의 공격 사정거리를 반환합니다.
         /// </summary>
         /// <returns>현재 그룹의 공격 사정거리.</returns>
-        public uint GetAttackRange()
+        public float GetAttackRange()
         {
             // TODO: 추후 유닛의 타입에 따라 사정거리 분기 기능 추가.
-            return 5;
+            return 5.0f;
         }
     }
 }
